+++ Plant-AR/Models/Favorites.swift
@@ -9,7 +9,6 @@ import Foundation
 import SwiftUI
 
 class Favorites: ObservableObject {
-
     @Published var plants: Set<String>
     let defaults = UserDefaults.standard
