+++ Plant-AR/Views/ContentView.swift
@@ -11,7 +11,6 @@ import ARKit
 import FocusEntity
 
 struct ContentView: View {
-
     @EnvironmentObject var arViewModel: ARViewModel
     @ObservedObject var favorites = Favorites()
     @State private var showFavorites = false
@@ -21,7 +20,6 @@ struct ContentView: View {
     @State private var placedModel: Model?
     @AppStorage("showIntro") var showIntro: Bool = true
     var loader = ModelLoader()
-
     var body: some View {
         NavigationView {
             ZStack(alignment: .bottom) {
