+++ Plant-AR/Views/MenuView/MenuButton.swift
@@ -10,7 +10,6 @@ import SwiftUI
 struct MenuButton: View {
     let symbol: String
     let action: () -> Void
-
     var body: some View {
         Button(action: {
             self.action()
