+++ Plant-AR/Views/MenuView/MenuView.swift
@@ -13,9 +13,7 @@ struct MenuView: View {
     @Binding var isInsertViewShown: Bool
     @Binding var selectedModel: Model?
     @EnvironmentObject var favorites: Favorites
-
     var models: [Model]
-
     var body: some View {
         VStack {
             Spacer()
