+++ Plant-AR/Views/PlantsView.swift
@@ -13,7 +13,6 @@ struct PlantsView: View {
     @Binding var selectedModel: Model?
     @EnvironmentObject var favorites: Favorites
     var models: [Model]
-
     var body: some View {
         NavigationView {
             ScrollView(showsIndicators: false) {
